+++ Modules/Intent.Modules.WindowsServiceHost/Templates/ProgramPartial/ProgramPartialTemplatePartial.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 using Intent.Engine;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
+using Intent.Modules.Common.CSharp.Events;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.CSharp.VisualStudio;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.VisualStudio.Projects.Api;
-using Intent.Modules.WindowsServiceHost.Events;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;